Fix for bug where float returned by API is unknowingly converted to a Double by the Entity class. Support for viewing the sentiment score of comments on site (but only when site is fully deployed),

+++ portfolio/src/main/java/com/google/sps/data/Comment.java
@@ -19,9 +19,9 @@ public final class Comment {
 
   private final String content;
   private final long timestamp;
-  private final float sentiment;
+  private final Double sentiment;
 
-  public Comment(String content, long timestamp, float sentiment) {
+  public Comment(String content, long timestamp, Double sentiment) {
     this.content = content;
     this.timestamp = timestamp;
     this.sentiment = sentiment;
